Rename Command.ts
