+++ src/discord/commands/command.d.ts
